create another wrapper for inasafe analysis utilities and qgis app internationalization
